Initial version and also replace of drop item
TODO: Go over this again

+++ code/game/machinery/bioprinter.dm
@@ -225,8 +225,9 @@
 			if(max_stored_matter == stored_matter)
 				to_chat(user, "<span class='warning'>\The [src] is too full.</span>")
 				return
+			if(!user.unEquipActive()) //If something prevents unequipping abort here
+				return
 			stored_matter += min(amount_list[path], max_stored_matter - stored_matter)
-			user.drop_item()
 			to_chat(user, "<span class='info'>\The [src] processes \the [W]. Levels of stored biomass now: [stored_matter]</span>")
 			qdel(W)
 			return
