+++ code/modules/mob/living/bot/ed209bot.dm
@@ -92,7 +92,7 @@
 	switch(build_step)
 		if(0, 1)
 			if(istype(W, /obj/item/robot_parts/l_leg) || istype(W, /obj/item/robot_parts/r_leg))
-				user.drop_item()
+				if(!user.unEquipActive()) return
 				qdel(W)
 				build_step++
 				to_chat(user, "<span class='notice'>You add the robot leg to [src].</span>")
@@ -110,7 +110,7 @@
 					if(!locate(/obj/item/clothing/accessory/armorplate) in W.contents)
 						to_chat(user, "There's no armor plates on this [W].")
 						return
-				user.drop_item()
+				if(!user.unEquipActive()) return
 				qdel(W)
 				build_step++
 				to_chat(user, "<span class='notice'>You add [W] to [src].</span>")
@@ -126,8 +126,7 @@
 					SetName("shielded frame assembly")
 					to_chat(user, "<span class='notice'>You welded the vest to [src].</span>")
 		if(4)
-			if(istype(W, /obj/item/clothing/head/helmet))
-				user.drop_item()
+			if(istype(W, /obj/item/clothing/head/helmet) && user.unEquipActive())
 				qdel(W)
 				build_step++
 				to_chat(user, "<span class='notice'>You add the helmet to [src].</span>")
@@ -136,8 +135,7 @@
 				icon_state = "ed209_hat"
 
 		if(5)
-			if(isprox(W))
-				user.drop_item()
+			if(isprox(W) && user.unEquipActive())
 				qdel(W)
 				build_step++
 				to_chat(user, "<span class='notice'>You add the prox sensor to [src].</span>")
@@ -160,13 +158,12 @@
 				return
 
 		if(7)
-			if(istype(W, /obj/item/weapon/gun/energy/taser))
+			if(istype(W, /obj/item/weapon/gun/energy/taser) && user.unEquipActive())
 				SetName("taser ED-209 assembly")
 				build_step++
 				to_chat(user, "<span class='notice'>You add [W] to [src].</span>")
 				item_state = "ed209_taser"
 				icon_state = "ed209_taser"
-				user.drop_item()
 				qdel(W)
 
 		if(8)
@@ -181,12 +178,11 @@
 					to_chat(user, "<span class='notice'>Taser gun attached.</span>")
 
 		if(9)
-			if(istype(W, /obj/item/weapon/cell))
+			if(istype(W, /obj/item/weapon/cell) && user.unEquipActive())
 				build_step++
 				to_chat(user, "<span class='notice'>You complete the ED-209.</span>")
 				var/turf/T = get_turf(src)
 				new /mob/living/bot/secbot/ed209(T,created_name,lasercolor)
-				user.drop_item()
 				qdel(W)
 				user.drop_from_inventory(src)
 				qdel(src)
